requirements.txt
streamlit==1.26.0
torch>=2.2.0
transformers==4.53.0
plotly==5.17.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.7
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
numpy==2.3.3
datasets==2.14.5
evaluate==0.4.6
accelerate==0.23.0
flask==3.0.0
streamlit
flask
flask-migrate
flask-sqlalchemy
datasets
argilla
python-dotenv==1.0.0
requests>=2.31.0